changes in all crud

+++ ALL_CRUD/MVC_CRUD/index.php
@@ -2,7 +2,7 @@
 <?php include(__DIR__ . '/../ADMIN/csslink.php'); ?>
       <script src="https://cdn.jsdelivr.net/npm/admin-lte@4.0.0-rc3/dist/js/adminlte.min.js" crossorigin="anonymous">
       </script>
-<?php include(__DIR__ . '/../ADMIN/sidebar.php'); ?>
+
 
         <?php
           ini_set('display_errors', 1);
@@ -13,11 +13,14 @@
           $controller = new UserController();
           $action = $_GET['action'] ?? 'index';
 
+          
           if ($action == 'store') $controller->store();
+        elseif ($action == 'add') $controller->edit($_GET['id']);
           elseif ($action == 'edit') $controller->edit($_GET['id']);
           elseif ($action == 'update') $controller->update($_GET['id']);
-          
           elseif ($action == 'delete') $controller->delete($_GET['id']);
           else $controller->index();
+
         ?>
-         <?php include(__DIR__ . '/../ADMIN/footer.php'); ?>
+        <?php include(__DIR__ . '/../ADMIN/sidebar.php'); ?>
+  <?php include(__DIR__ . '/../ADMIN/footer.php'); ?>
